LPS-55583 Moves java
