+++ modules/apps/user-groups-admin/user-groups-admin-web/src/com/liferay/user/groups/admin/web/lar/UserGroupsAdminPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.usergroupsadmin.lar;
+package com.liferay.user.groups.admin.web.lar;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.lar.BasePortletDataHandler;
