+++ modules/apps/user-groups-admin/user-groups-admin-web/src/com/liferay/user/groups/admin/web/util/UserGroupIndexer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.usergroupsadmin.util;
+package com.liferay.user.groups.admin.web.util;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.exception.PortalException;
